Rename `osu.Framework.Platform.SDL2` to `osu.Framework.Platform.SDL`
`.SDL3` has a naming conflict with `SDL3` static class
